+++ src/models/state.ts
@@ -1,9 +1,7 @@
-import { ExtensionContext } from "vscode";
 import { ISyncer } from "~/models";
 import { Watcher } from "~/services";
 
 export interface IExtensionState {
-  context?: ExtensionContext;
   sync?: ISyncer;
   watcher?: Watcher;
 }